requirements.txt
numpy
pandas
plotly
pydeck
streamlit
folium
branca.colormap
